11.3 Watch: find your terminal and run:   php bin/console importmap:remove lodash  That will remove lodash from importmap.php and delete the file from assets/vendor/.
 Nice! Now run ./bin/console importmap:require with the package name / the path that you want: lodash/camelCase.js.

  php bin/console importmap:require lodash/camelCase.js
camelCase.js is the name of the file over on the CDN. But you'll notice that, a lot of times in the docs, they'll reference lodash/camelCase without the .js. And in this case, you can leave the .js off: it's up to you. That works because jsDelivr is friendly and makes both versions of the URL work.

 The result of the command? The same as before! We get a new entry in importmap.php matching what we want to import and set to a URL.

 Copy that URL so we can see it. There we go! It's the code from just camelCase.js.

 And when we try the page... it works!

 Here's the takeaway: if you need to import a specific file from a package, you can do that: just pass the package name + file path to importmap:require.

 Next, let's add Stimulus to our app!