20.5 No big deal. This is just another thing we need to add to our build step... before we run asset-map:compile so that the file is available.
I'll paste in the code for this. This is basically the same code we ran earlier to set things up locally, except that we're using the linux-x64 build. We're downloading that, moving it into the /bin directory (it doesn't really matter where it goes), making it executable, and then running that same command so that the output file is there by the time asset-map:compile runs.

 Oh, and don't forget about the TailwindBundle which makes using Tailwind - including this deploy step - a bit easier.

 Back over here, let's commit that new config change.... then deploy again:

  symfony deploy
 Even while it's deploying, we can see that this working. Last time, there were 16 files, now there are 17. When it finishes, spin over, refresh and... it's alive! All the pages have CSS. I love it!

 Now that we're on production, let's talk about the things we need to check to make sure our assets are served blazingly fast.